Photos/Version-2/Framework/requirements.txt
# HSTL Photo Framework Dependencies
# Version: 1.8.0
# Python Requirement: 3.9+

# ============================================================================
# CORE FRAMEWORK DEPENDENCIES
# ============================================================================

# Configuration parsing
PyYAML>=6.0                    # YAML configuration file parsing

# Data manipulation and Excel/CSV processing
pandas>=2.0.0                  # Data manipulation, Excel reading, CSV export
openpyxl>=3.0.0                # Required by pandas for .xlsx file reading

# Text encoding (Step 3: Unicode Filtering)
ftfy>=6.0.0                    # Mojibake detection and text encoding repair

# GUI framework
PyQt6>=6.4.0                   # Main GUI framework
pyqt-app-info @ git+https://github.com/juren53/pyqt-app-info.git  # About dialog and app info

# Image processing (Step 8: Watermark Addition)
Pillow>=10.0.0                 # Image manipulation for watermark compositing
PyExifTool>=0.5.0              # Read/write IPTC metadata, copy metadata to output

# ============================================================================
# OPTIONAL DEPENDENCIES
# ============================================================================

# These are used by auxiliary tools, not the main HPM application

# wxPython>=4.2.0              # Only needed for csv_record_viewer.py (standalone tool)

# ============================================================================
# DEVELOPMENT AND TESTING DEPENDENCIES
# ============================================================================

# Uncomment these for development work
# pytest>=7.0.0                # Unit testing framework
# pytest-cov>=4.0.0            # Test coverage reporting

# ============================================================================
# NOTES
# ============================================================================

# 1. Minimum install for HPM GUI:
#    pip install PyYAML pandas openpyxl ftfy PyQt6 Pillow PyExifTool
#
# 2. The following packages listed in previous versions are NOT currently used:
#    - tqdm (progress bars) - Not imported anywhere
#    - structlog (logging) - Not imported anywhere
#    - colorama - Not imported anywhere
#    - pydantic - Not imported anywhere
#    - xlrd - Not imported anywhere (pandas uses openpyxl for .xlsx)
#
# 3. Google API libraries (google-api-python-client, gspread, etc.) are only
#    used in archived code and are not required for current HPM functionality.
#
# 4. ExifTool command-line tool is required (separate install):
#    PyExifTool is a Python wrapper - the actual ExifTool executable must be
#    installed separately. Download from: https://exiftool.org/
#    On Windows, rename exiftool(-k).exe to exiftool.exe and add to PATH.
#
# 5. Standard library modules used (no pip install needed):
#    pathlib, typing, dataclasses, datetime, logging, os, sys, subprocess,
#    shutil, json, csv, re, threading, time, enum, urllib, glob, platform
